+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Moved docker-compose images from docker.io to mirror.gcr.io (OSIDB-3653)
+- Removed `last_validated_dt` from exposed JSON Flaw History data (OSIDB-3814)
 
 ## [4.6.3] - 2025-01-09
 ### Fixed